Microsteps: Adjusted offset selection to work with steep microsteps
When encoding a waveform with 3 current steps per step surrounded by
current steps of less than two, the correct encoding is an offset 2 +
bits 1 encoding for the steep part. The previous sections calculation
would try to generate an offset 3 + bits 0 encoding, which is invalid
due to a too large offset.

Effectively the old implementation favored having the bits set to 0 in
section. Now the implementation favors offsets closer to 0.